weaviate-demo/requirements.txt
weaviate-client==4.9.6
sentence-transformers==3.3.1
python-dotenv==1.0.1
